feat: useVisible 훅 분리
